pyproject.toml
[project]
name = "pynodewidget"
version = "0.1.0"
description = "A Python wrapper for ReactFlow using AnyWidget - build interactive node graphs without JavaScript"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anywidget>=0.9.0",
    "pydantic>=2.12.4",
    "traitlets>=5.14.0",
]
authors = [
    { name = "Henning", email = "henning.scheufler@web.de" }
]
keywords = ["jupyter", "widget", "reactflow", "nodes", "graph"]
classifiers = [
    "Framework :: Jupyter",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
]

license = "MIT"
license-files = ["LICENSE"]

[project.urls]
Homepage = "https://henningscheufler.github.io/pynodewidget/"
Issues = "https://github.com/HenningScheufler/pynodewidget/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "taskipy>=1.13.0",
    "hatchling>=1.27.0",
    "mypy>=1.0.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
    "pymdown-extensions>=10.8.0",
]
notebooks = [
    "jupyterlab>=4.5.0",
    "marimo>=0.18.0",
]
all = [
    "pynodewidget[dev,docs,notebooks]",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pynodewidget"]

[tool.hatch.build.targets.wheel.hooks.custom]
path = "hatch_build.py"

[tool.hatch.build.targets.sdist]
[tool.hatch.build.targets.sdist.hooks.custom]
path = "hatch_build.py"


[tool.taskipy.tasks]
docs-serve = "mkdocs serve"
docs-build = "mkdocs build"
test-py = "pytest"
test-js = "bun --cwd=js run test"
test-py-cov = "pytest --cov=pynodewidget --cov-report=xml:coverage.xml --cov-report=html:coverage_html"
test-js-cov = "bun --cwd=js run test:coverage"
test = "task test-py && task test-js"


[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers -q"

[tool.coverage.run]
omit = [
    "tests/*",
    "docs/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == '__main__':",
]

[tool.coverage.html]
directory = "coverage_html"

